:bee: 把Features Function數組移動到單獨文件
